// _audit/BUDCAST_FULL_AUDIT_BUILD_2026-04-29/packages/shared/src/types/database.ts
export type UserType = 'creator' | 'brand' | 'brand_team';
// ...
export type Badge =
  | 'verified_brand'
  | 'verified_creator'
  | 'payment_ready'
  | 'campaign_ready'
  | 'social_verified'
  | 'verified_payer'
  | 'trusted_creator'
  | 'highly_rated'
  | 'top_rated'
  | 'founding_brand'
  | 'founding_creator';